Update Collidable to use doubles

+++ src/main/java/ca/bcit/comp2522/termproject/jaguarundi/Collidable.java
@@ -1,8 +1,8 @@
 package ca.bcit.comp2522.termproject.jaguarundi;
 
 public interface Collidable {
-    int getXPosition();
-    int getYPosition();
-    int getWidth();
-    int getHeight();
+    double getXPosition();
+    double getYPosition();
+    double getWidth();
+    double getHeight();
 }
